backend_service/src/video_pipeline/requirements.txt
boto3
requests
ffmpeg-python
moviepy
Pillow
requests>=2.28.0
